feat(batch): déduction du stock + portions/conservation réelles à la cuisson
Le batch cooking ne déduisait RIEN du stock — il créait juste le plat
cuisiné avec les portions du plan. Désormais la cuisson d'un batch peut
déduire les ingrédients réellement utilisés (FEFO) et fixer les portions
faites + le mode de conservation.

- lib/deductNeeds.js : helper partagé (besoins FEFO multi-lots +
  déductions explicites), extrait de /api/meals/cook (refactor sans
  changement de comportement — 144 tests OK) et réutilisé par le batch
- POST /api/planning/batch/cook accepte needs[]/portions/storage_method ;
  déduit le stock, crée le cooked_dish avec les vraies portions + DLC
  (frigo: keeps_days du plan ; congélateur: 90 j) ; renvoie shortfalls.
  Rétro-compatible (sans needs[] = comportement inchangé)
- GET /api/planning/batch/[batchRecipeId]/ingredients : parse + résout le
  texte d'ingrédients du batch pour pré-remplir la feuille de cuisson

L'UI (mode cuisson batch dans CookSession + câblage page batch) suit dans
un commit séparé.
