+++ app/api/meals/cook/route.js
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server'
 import { authenticateRequest } from '@/lib/apiAuth'
 import { calculateCookedDishExpiration } from '@/lib/shelfLifeRules'
-import { allocateConsumption } from '@/lib/stockAllocator'
+import { deductFromStock } from '@/lib/deductNeeds'
 
 export const dynamic = 'force-dynamic'
 
@@ -168,81 +168,13 @@ export async function POST(request) {
     }
   }
 
-  // 2a. Besoins « automatiques » : { canonical_food_id | archetype_id, qty, unit }
-  //     → répartition multi-lots (lot ouvert d'abord, puis FEFO, un seul lot
-  //     entamé) via lib/stockAllocator ; le trigger SQL auto_open_lot marque
-  //     « ouvert » le lot partiellement consommé et réduit sa DLC.
-  //     Ex : 5 bouteilles de 1 L, besoin 1,5 L → 1 vidée + 0,5 L sur la 2e.
-  const allDeductions = [...(deductions || [])]
-  const shortfalls = []
-  for (const n of (needs || [])) {
-    const qty = Number(n?.qty)
-    if (!(qty > 0) || !n?.unit || (!n?.canonical_food_id && !n?.archetype_id)) continue
-
-    let lotQuery = supabase
-      .from('inventory_lots_resolved')
-      .select('id, qty_remaining, unit, expiration_date, user_id')
-      .eq('user_id', user.id)
-      .gt('qty_remaining', 0)
-    lotQuery = n.canonical_food_id
-      ? lotQuery.eq('resolved_canonical_food_id', n.canonical_food_id)
-      : lotQuery.eq('resolved_archetype_id', n.archetype_id)
-    const { data: candidateLots } = await lotQuery
-
-    if (!candidateLots?.length) {
-      shortfalls.push({ ...n, missing: qty })
-      continue
-    }
-
-    // is_opened / adjusted ne sont pas exposés par la vue : on les recharge
-    const ids = candidateLots.map(l => l.id)
-    const { data: fullLots } = await supabase
-      .from('inventory_lots')
-      .select('id, qty_remaining, unit, is_opened, expiration_date, adjusted_expiration_date')
-      .in('id', ids)
-
-    let meta = {}
-    if (n.canonical_food_id) {
-      const { data: cf } = await supabase
-        .from('canonical_foods')
-        .select('unit_weight_grams, density_g_per_ml')
-        .eq('id', n.canonical_food_id)
-        .maybeSingle()
-      meta = { grams_per_unit: cf?.unit_weight_grams, density_g_per_ml: cf?.density_g_per_ml }
-    }
-
-    const { allocations, shortfall } = allocateConsumption(fullLots || [], qty, n.unit, meta)
-    for (const a of allocations) {
-      allDeductions.push({ lot_id: a.lot_id, quantity_used: a.qty_in_lot_unit })
-    }
-    if (shortfall > 0) shortfalls.push({ ...n, missing: shortfall })
-  }
-
-  // 2b. Déduction du stock (on garde les DLC des lots pour la DLC du reste éventuel)
-  let deductedCount = 0
-  const usedLots = []
-  for (const d of allDeductions) {
-    const lotId = d?.lot_id
-    const qty = Number(d?.quantity_used)
-    if (!lotId || !(qty > 0)) continue
-    const { data: lot, error: lotErr } = await supabase
-      .from('inventory_lots')
-      .select('id, qty_remaining, adjusted_expiration_date, expiration_date, best_before')
-      .eq('id', lotId)
-      .eq('user_id', user.id)
-      .single()
-    if (lotErr || !lot) continue
-    const newQty = Math.max(0, (lot.qty_remaining || 0) - qty)
-    const { error: updErr } = await supabase
-      .from('inventory_lots')
-      .update({ qty_remaining: newQty })
-      .eq('id', lotId)
-      .eq('user_id', user.id)
-    if (!updErr) {
-      deductedCount++
-      usedLots.push(lot)
-    }
-  }
+  // 2. Déduction du stock : besoins automatiques (FEFO multi-lots) + déductions
+  //    explicites, via le helper partagé (cf. lib/deductNeeds). Le trigger SQL
+  //    auto_open_lot marque « ouvert » un lot partiellement consommé.
+  //    Ex : 5 bouteilles de 1 L, besoin 1,5 L → 1 vidée + 0,5 L sur la 2e.
+  const { deductedCount, usedLots, shortfalls } = await deductFromStock(
+    supabase, user.id, { needs, deductions },
+  )
 
   // 3. Décrément du plat consommé (reste mangé ou portion de batch).
   //    Une seule fois par créneau (cf. JSDoc).
